refactor: prevent useless re-rendering
