--- src/shared/hook/index.ts
@@ -1,1 +1,0 @@
-export { useViewport } from './useViewport';
