+++ src/shared/store/index.ts
@@ -1,1 +1,2 @@
 export type { DeviceType, DeviceTypeState } from './Device';
+export { useDeviceStore } from './useDeviceStore';
